DataBaseLogger & FileLogger Added

+++ src/main/java/com/example/rentacarv1/repositories/LogRepository.java
@@ -3,6 +3,6 @@ package com.example.rentacarv1.repositories;
 import com.example.rentacarv1.entities.concretes.Log;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LogRepository extends JpaRepository<Log,Integer> {
+public interface LogRepository extends JpaRepository<Log,Long> {
 
 }
